--- e2e/fixtures/citizen/caseView/respondent/viewAllDocuments/respondentAllCategoriesContent.ts
@@ -1,8 +1,0 @@
-export enum RespondentAllCategoriesContent {
-  pageTitle = "View all documents",
-  p = "Case number",
-  p1 = "Categories",
-  p2 = "Last update",
-  a1 = "Your served application pack",
-  a2 = "Applicant's documents",
-}
