--- e2e/fixtures/citizen/caseView/respondent/viewAllDocuments/respondentApplicationPackDocumentsContent.ts
@@ -1,11 +1,0 @@
-export enum RespondentApplicationPackDocumentsContent {
-  pageTitle = "Your served application pack",
-  span = "View all documents",
-  p1 = "Case number", //add case number
-  p2 = "You should read the cover letter first as this tells you what to do next. The cover letter also gives you more information on the other documents in your pack and what you need to do with them.",
-  a1 = "coversheet.pdf",
-  a2 = "FL401FinalDocument.pdf",
-  a3 = "Privacy_Notice.pdf",
-  a4 = "Power_of_arrest.pdf",
-  a5 = "mockFile.pdf",
-}
