+++ e2e/fixtures/citizen/caseView/viewAllDocuments/allCatergoriesContent.ts
