+++ e2e/fixtures/citizen/caseView/viewAllDocuments/applicationPackDocumentsContent.ts
