+++ e2e/journeys/citizen/caseView/viewAllDocuments/viewAllDocuments.ts
@@ -1,7 +1,7 @@
 import { ActivateCase, CaseUser } from "../../activateCase/activateCase.ts";
 import { Browser, Page } from "@playwright/test";
-import { AllCategoriesPage } from "../../../../pages/citizen/caseView/viewAllDocuments/applicant/allCatergoriesPage.ts";
-import { ApplicationPackDocumentsPage } from "../../../../pages/citizen/caseView/viewAllDocuments/applicant/applicationPackDocumentsPage.ts";
+import { AllCategoriesPage } from "../../../../pages/citizen/caseView/viewAllDocuments/allCatergoriesPage.ts";
+import { ApplicationPackDocumentsPage } from "../../../../pages/citizen/caseView/viewAllDocuments/applicationPackDocumentsPage.ts";
 
 interface reasonableAdjustmentsParams {
   page: Page;
