+++ e2e/pages/citizen/caseView/viewAllDocuments/applicationPackDocumentsPage.ts
@@ -1,15 +1,15 @@
-import { RespondentApplicationPackDocumentsContent } from "../../../../../fixtures/citizen/caseView/respondent/viewAllDocuments/respondentApplicationPackDocumentsContent.ts";
-import { Selectors } from "../../../../../common/selectors.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
-import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper.ts";
+import AccessibilityTestHelper from "../../../../common/accessibilityTestHelper.ts";
+import { ApplicationPackDocumentsContent } from "../../../../fixtures/citizen/caseView/viewAllDocuments/applicationPackDocumentsContent.ts";
+import { Selectors } from "../../../../common/selectors.ts";
 import { Page } from "@playwright/test";
+import { Helpers } from "../../../../common/helpers.ts";
 
 interface ApplicationPackDocumentsParams {
   page: Page;
   accessibilityTest: boolean;
 }
 
-export class RespondentApplicationPackDocumentsPage {
+export class ApplicationPackDocumentsPage {
   public static async applicationPackDocumentsPage({
     page,
     accessibilityTest,
@@ -27,29 +27,29 @@ export class RespondentApplicationPackDocumentsPage {
     }
     await page
       .locator(Selectors.h1, {
-        hasText: RespondentApplicationPackDocumentsContent.pageTitle,
+        hasText: ApplicationPackDocumentsContent.pageTitle,
       })
       .waitFor();
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.Span}:text-is("${RespondentApplicationPackDocumentsContent.span}")`,
+        `${Selectors.Span}:text-is("${ApplicationPackDocumentsContent.span}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.p}:has-text("${RespondentApplicationPackDocumentsContent.p1}")`,
+        `${Selectors.p}:has-text("${ApplicationPackDocumentsContent.p1}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.p}:has-text("${RespondentApplicationPackDocumentsContent.p2}")`,
+        `${Selectors.p}:has-text("${ApplicationPackDocumentsContent.p2}")`,
         1,
       ),
       Helpers.checkGroup(
         page,
         5,
-        RespondentApplicationPackDocumentsContent,
+        ApplicationPackDocumentsContent,
         "a",
         Selectors.a,
       ),
@@ -66,7 +66,7 @@ export class RespondentApplicationPackDocumentsPage {
       throw new Error("Page is not defined)");
     }
     await page.click(
-      `${Selectors.a}:text-is("${RespondentApplicationPackDocumentsContent.a1}")`,
+      `${Selectors.a}:text-is("${ApplicationPackDocumentsContent.a1}")`,
     );
     await page.waitForTimeout(2000);
   }
